Refactor temporary contact model
