+++ core/jobs.py
@@ -6,8 +6,8 @@ from django_task.job import Job
 from django.core.mail import send_mail
 
 from core import kronos
+from core.models import TemporaryContact
 from core.parsers import KronosEventsParser, KronosParticipantsParser
-from core.temp_models import create_temporary_table, TemporaryContact, db_table_exists
 from core.utils import ConflictResolutionMethods, update_object
 
 
@@ -56,7 +56,6 @@ class LoadKronosParticipants(Job):
             print(event_ids)
             response = kronos_client.get_participants(event_ids)
             print(response)
-            create_temporary_table()
             parser = KronosParticipantsParser(task=task)
             parser.parse_contact_list(response.get("records"))
             task.description = (
@@ -77,33 +76,28 @@ class ResolveAllConflicts(Job):
     def execute(job, task):
         task.log(logging.INFO, "Resolving conflicts")
         try:
-            if db_table_exists("core_temporarycontact"):
-                if task.method == ConflictResolutionMethods.KEEP_OLD_DATA:
-                    cursor = connection.cursor()
-                    cursor.execute("DROP TABLE core_temporarycontact")
-                elif task.method == ConflictResolutionMethods.SAVE_INCOMING_DATA:
-                    incoming_contacts = TemporaryContact.objects.select_related(
-                        "record"
-                    ).all()
-                    for incoming_contact in incoming_contacts:
-                        try:
-                            task.log(
-                                logging.INFO,
-                                f"Resolving conflict for {incoming_contact}",
-                            )
-                            record = incoming_contact.record
-                            update_values = copy(vars(incoming_contact))
-                            update_values.pop("record_id")
-                            update_values.pop("id")
-                            update_object(record, update_values)
-                            TemporaryContact.objects.filter(
-                                pk=incoming_contact.id
-                            ).first().delete()
-                        except Exception as e:
-                            task.log(logging.ERROR, e)
-                    if TemporaryContact.objects.count() == 0:
-                        cursor = connection.cursor()
-                        cursor.execute("DROP TABLE core_temporarycontact")
+            if task.method == ConflictResolutionMethods.KEEP_OLD_DATA:
+                TemporaryContact.objects.all().delete()
+            elif task.method == ConflictResolutionMethods.SAVE_INCOMING_DATA:
+                incoming_contacts = TemporaryContact.objects.select_related(
+                    "record"
+                ).all()
+                for incoming_contact in incoming_contacts:
+                    try:
+                        task.log(
+                            logging.INFO,
+                            f"Resolving conflict for {incoming_contact}",
+                        )
+                        record = incoming_contact.record
+                        update_values = copy(vars(incoming_contact))
+                        update_values.pop("record_id")
+                        update_values.pop("id")
+                        update_object(record, update_values)
+                        TemporaryContact.objects.filter(
+                            pk=incoming_contact.id
+                        ).first().delete()
+                    except Exception as e:
+                        task.log(logging.ERROR, e)
 
         except Exception as e:
             task.log(logging.ERROR, e)
